chore: useDateRange hook
